requirements.txt
fastapi
uvicorn[standard]
python-dateutil
pydantic
google-generativeai
